Added all new tables and seed data to DbContext; save point before update migration

+++ Models/NeedleLength.cs
@@ -13,7 +13,7 @@ namespace StashMash.Models
 
         [Required(ErrorMessage = "Please enter the length of the needle.")]
         [Range(1, 5, ErrorMessage = "The length cannot exceed 5 digits.")]
-        public decimal Length { get; set; }
+        public double Length { get; set; }
 
         public int MeasurementUnitId { get; set; }
         public MeasurementUnit MeasurementUnit { get; set; }
